(PC-27670)[PRO] refactor: Remove unnecessary file sendSentryError.
